refactor: Login, SignUP request가 상속하는 클래스를 생성하여 Service method 매개변수 통일
