+++ backend/footprint/src/main/java/foot/footprint/domain/member/dto/authRequest/LoginRequest.java
@@ -1,0 +1,11 @@
+package foot.footprint.domain.member.dto.authRequest;
+
+import lombok.Getter;
+
+@Getter
+public class LoginRequest extends AuthRequest {
+
+    public LoginRequest(String email, String password) {
+        super(email, password);
+    }
+}
