+++ backend/footprint/src/test/java/foot/footprint/service/member/AuthServiceTest.java
@@ -5,8 +5,9 @@ import foot.footprint.domain.member.dao.MemberRepository;
 import foot.footprint.domain.member.domain.AuthProvider;
 import foot.footprint.domain.member.domain.Role;
 import foot.footprint.domain.member.domain.Member;
-import foot.footprint.domain.member.dto.LoginRequest;
-import foot.footprint.domain.member.dto.SignUpRequest;
+import foot.footprint.domain.member.dto.authRequest.AuthRequest;
+import foot.footprint.domain.member.dto.authRequest.LoginRequest;
+import foot.footprint.domain.member.dto.authRequest.SignUpRequest;
 import foot.footprint.domain.member.exception.AlreadyExistedEmailException;
 import foot.footprint.domain.member.exception.NotMatchPasswordException;
 import foot.footprint.global.error.exception.NotExistsException;
